Install make to build wheels
We add make on platform dpkg to our bindep list in order to ensure it is
there when we want to build wheels. In particular this is necessary for
container image builds on arm64 as it isn't already there in the base
image.

Bump image build timeout as arm64 image takes more time to build.

Change-Id: I3b358721eebbceafc12daf9d706306634048b196