Omnibus nodepool container image fixes
We updated python-base and python-builder to include arm64 images in
support of nodepool's arm64 python-builder image. In doing so we have
discovered a number of issues that all need to be addressed together to
get a working arm64 nodepool-builder image.

First up we need to install make via bindep because unlike x86_64 pypi
does't have prebuild wheels for arm64 on many packages. That means we
have to build them. After fixing that we've discovered that the wheel
builds under qemu with buildx are not very quick. In order to speed that
up we consume the OpenDev linaro cloud arm64 wheel cache. This doesn't
have wheels for every package we need (so the make fix stays), but it
does have them for the most costly builds: pynacl, cryptography, bcrypt.

With wheels fixed we discovered we needed vhd-util for arm64 on the
nodepool-builder image. This has been consumed from the openstack infra
PPA for x86 and was updated to do ubuntu focal builds for arm64 as well.
We can consume them with a minor update to the PPA config.

Finally we increase the job timeout to an hour as the buildx stuff is
still a bit slower than x86.

Co-Authored-By: James E. Blair <corvus@inaugust.com>
Co-Authored-By: Ian Wienand <iwienand@redhat.com>
Change-Id: I3b358721eebbceafc12daf9d706306634048b196