Add basic max-servers handling to the k8s driver
This maps zk "nodes" to k8s resources which are currently pods or
namespaces in order to count them and apply max-servers limitations.
Since a pod is always in a namespace we've essentially mapped "nodes"
to k8s namespaces. This limitation is quite basic and jobs with a
namespace can still in theory consume a number of resources, but if you
control the jobs and then limit the number of concurrent jobs via
max-servers aka max namespaces you should be able to have some control.

The code is stubbed out such that a followon can query k8s for
additional limitations and try to apply them more properly.

Change-Id: Ieb9f179d99b322a9cbf4bc19d4f495bcda0d1ea8