Check for images to upload single threaded
Currently we have multiple upload workers where each of them is
constantly checking if there is an image to upload. This check is
involves quite a lot zk operations increasing load on the
nodepool-builder as well as on zk. Instead we can run the check loop
only once and dispatch the actual upload to a limited
ThreadPoolExecutor.

Change-Id: I95a4a57c4acb856de5c25d570249c43f735503d5