Update patch set 24
Patch Set 24: Code-Review-1

This seems like it's at least partially re-implementing the TaskManager that we just added.  That's a way of handling this sort of thing that all drivers can share.

Moreover, the "simple" driver interface uses that, and I *think* we should be able to port the AWS driver to it.

Can you take a look at that and see if (a) the AWS driver could be ported to the simple driver?  and if not, (b) can this use the TaskManager?

Patch-set: 24
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1