Update patch set 25
Patch Set 25:

Yes it could be great to implement one of these methods in this provider and there's also some other work to do in this provider.
We are looking to allocate time on our side for this but for now we were just looking for a patch for the rate limiting issue.
It's causing a lot of failures for some reviews on our side.
Would be it ok to merge this review now, knowing we will work on the provider later ?

Patch-set: 25
Label: Verified=0