Load diskimage configs before building them
Currently we load our diskimage configs then iterate over all of them
rebuilding them if necessary. The problem is that since disk image
builds take so long we don't pick up pause configurations until we finish
attempting to build all images. Fix this by checking if our config
is current before building images. If not, return and force a rerun
through the rebuilding process which will load new configs.

This also fixes a test race that was noted in comments but a non issue
until this change. This change causes us to trip over the test race
frequently.

Change-Id: If0b7321ee2c804a4ae8123f2369a660780272b52