Speed up node list
When determining the locked status of a node the node list tries to lock
and immediately unlocks. When having lots of nodes in the system this
takes very long and puts a considerable amount of load onto zookeeper.
Since the lock information in the node list doesn't really need to lock
nodes leverage the already existing tree cache to just check for lock
contenders in the cache.

Change-Id: I6a1760315a13cadcbaf1fbd8e7be202bc9c5b32a