Silence paramiko exceptions for incompatible keys
Now that we iterate over all available SSH host keys, there is no
need to raise an exception when it can't retrieve at least one of
them. This seems to have been the original intent of change
Ibb2a07a29681dcefd4017eb2fd6134ee33ab726c but it lacked a pass
condition for this case.

Change-Id: I45b46638d11da4a24fc562158bc6e779352a71d9