Add images-dir-required-free option
A builder close to a full disk has a guaranteed death sentence; it
will start a build that can't complete, leaving random builds around
until the disk reaches 100%, and then goes into a loop trying to run
builds that will never work and rotating out quickly any useful build
logs that might have helped debug if there was an issue with image
size.

This adds a global and diskimage-specific option
images-dir-required-free which checks the free space on the configured
images-dir volume and will not start a build if it is less than
required.

Operators who know the size of their images and have knowledge of
things like the ephemeral space required for image conversions,
etc. can set this limit to something appropriate and avoid courting a
failure.  If another builder is available, it should be able to pick
up the required builds, sharing the load more effectively.

Change-Id: I5cca2b64faa5455d0eb9d2501cfe702cc15db637