Bump openshift dep
Note this isn't expected to pass testing, but this change is being used
to start to understand what changes are necessary to nodepool to support
an up to date openshift library.

The translation of old openshift library calls to currently supported
and available calls has begun but is not complete. We also likely need
to update our fakes in testing.

Change-Id: Ia4d09fd0a4a49d644bb575b74184de930c62ce89
Story: 2008427
Task: 41373