Use libyaml parsing when available
Nodepool periodically reloads its config. On systems with large
configs this can cause a substantial amount of cpu usage. Our
profiling of an idle provider revealed that the yaml parsing accounted
for ~15% cpu load while the provider was idle.

This could be improved by utilizing the native libyaml bindings when
available.

Change-Id: Icbed393c5aa00abd52d1e13ccbbad8546ec6cf97