Add ssh/winrm credentials to static nodes
This allows associating specific ssh private key files or winrm
certificates with static nodes.  Nodepool will not use the files
directly, but instead will add their contents to the ZooKeeper
node record for the nodes.  Zuul can then read these values and
use them in preference to the key files supplied with the executor
configuration.

The credential record is designed to be extensible to support other
systems (eg, kerberos) in the future if desired.

Eventually, a similar feature could be added to the dynamic Nodepool
drivers so that Zuul doesn't need to have private keys (and we could
stop splitting the credential information across both systems which
should ease deployment).

Change-Id: I1d0cb5579326da258e94b916dd666aa82af3582e