Optimize list_servers call in cleanupLeakedInstances
By default list_servers gets the server list and then attaches the
network information to each server object. This involves a call to the
floatingips endpoint per instance. This is unneeded during
cleanupLeakedInstances since we're only interested in the id and
metadata in order to trigger node deletion. In larger clouds this can
trigger hundreds or thousands unnecessary api calls to the
cloud. Optimize this by getting the server list with the bare flag to
avoid that.

Change-Id: Ie48e647eefc2aa3169e943fdf6854d82219b645b