Add logging to noderequest handler polls
We have observed the occasional situation where a noderequest will have
all of its nodes built and ready, but the request is never marked
fulfilled. This results in nodes that are ready and locked and won't be
used by zuul. The zuul builds waiting for these nodes wait indefinitely
as a result.

I suspect that the noderequest handler poll is returning early which
prevents it from ever unlocking the nodes and marking the request
fulfilled. Add some logging to locations where the poll can return
early.

Note this may be too verbose and we should double check what this looks
like in test logs before deciding to merge it.

Change-Id: If4c0a4b0794c73fb995c1de6c612ee16929d2b24