Update patch set 3
Patch Set 3:

Couple of questions:

1) Should we include support for ansible_shell_executable at the same time?  I realize this handles the default case.  Maybe that's all we really need.
2) Why not Azure?  I see that you updated the test to assert that it's not set, so it doesn't seem like an accident.  I don't see a reason to omit it, so I'm thinking this might be something we need to fix.

Bonus question just to validate my understanding: I assume this works the same way with a kubectl connection type if we're using k8s pods with the ansible shell/command module.  And there's no effect if we use Nodepool to provide namespaces.  Is that right?  (I asked Albin this in IRC too.)

Patch-set: 3
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>