Improve state machine keyscan handling
This is two related changes to the statemanager driver in one:

Add a dedicated thread to independency drive state machines.
Currently, the state machines are driven by each provider-pool's
pool worker thread, which means that there is a 10 second sleep
between each run of the loop.  That's good for the pool worker
because it reduces ZK traffic, but it's too slow for driving a
bunch of parallel state machines.  The new thread is started by
the provider manager and simply advances all of the state machines
for that provider manager.

We also start a ThreadPoolExecutor to perform keyscans.  Previously
they would happen synchronously and serially in the pool worker
thread, which means that when a large number of nodes simultaneously
finished, we would slowly keyscan each one.  Now the new thread above
will dispatch keyscan requests to the pool, and when they are
complete, will mark the launch finished.

Assuming this is mostly IO-bound for now, a ThreadPool should be
sufficient; however if we later find that we want to escape the GIL,
we can switch to a ProcessPool.  A thread pool, however, has fewer
potential issues with process management (oom, etc), so should be
preferred for now.

Change-Id: Ibd8d109969c1eaa898b1a7bca2740253dbc17c73