Convert NodeLaunchRecord into NodeLauncher
In current nodepool drivers, a NodeLauncher is a thread; the main
objective of the statemachine driver was to eliminate that thread.
But we still need to track data about an individual node launch.
Let's use a class that implements the same API as the NodeLauncher
class, but is not a thread.  We'll manage it in the same way as
we have been managing the state machines, but by implementing this
API, we can use helper methods like the stats reporter.

Ultimately, the concept is the same: an object that a request handler
uses to effectuate and track a node launch.

Change-Id: I0b6aa2077740ba788f14837893ab08213802802b