Uniquely identify launchers
Previously launchers were identify as:

  hostname-pid-provider-pool

The problem with this setup is that using containers with host
networking and namespaced pids you can end up with multiple launchers
using the same launcher id for the same provider pool. This problem also
arises if you are replacing launcher1.example.com with
launcher1.otherexample.com.

To fix this we update the launcher ids to be:

  fqdn-randomuuid-provider-pool

These ids are already not expected to survive beyond the lifetime of any
single process as they use a pid value. This means we can replace this
pid value with a randomly generated uuid value instead. We also use the
host fqdn instead of hostname to be less ambiguous in the case where two
different hosts with different fqdns share a hostname.

Change-Id: I419718e63b31b12d8dfe971031cd8a81ad582480