Azure: delete disk based on VM reference
It can take up to 30 minutes after creation for a disk to appear
in the list of disks returned by the API.  This means that if we
rely on querying the API to find the disk to delete when deleting
a VM, we will not see it if the VM was used for less than 30m.  In
that case, we would always rely on the leaked resource cleanup to
delete the disk, and disks would stay in use longer than necessary.

To correct this, save the os disk name when we start deleting the
VM, and use that to delete the disk.

Change-Id: I97756165c426371833c6d014125fce341dad960e