Azure: reconcile config objects
The config objects in the Azure driver have drifted a bit.  This
updates them to match the actual used configuration.  It also
reorganizes them to be a little easier to maintain by moving the
initializers into the individual objects.

Finally, the verbose __eq__ methods are removed in favor of a
simpler __eq__ method in the superclass.

Since the OpenStack driver calls super() in its __eq__ methods,
it needs to be updated at the same time.

Change-Id: Id6971ca002879d3fb056fedc7e4ca6ec35dd7434