Fix possible thread leak
There is a possible race when stopping a provider (which can happen
during config updates) that self.running is set to False and the
server_list watcher thread exits without waking up the waiting
threads. To fix this remove the running flag and just rely on the stop
event to exit the loop.

Change-Id: Ib59f9038149aa5c5a46ca2a2a1a0f22795935ab4