Fix possible thread leak
There is a possible race when stopping a provider (which can happen
during config updates) that self.running is set to False and the
server_list watcher thread exits without waking up the waiting
threads. To fix this remove the running flag and just rely on the stop
event to exit the loop. Further to keep the optimized version stop the
watcher thread during join after the waiting queues are empty.

Change-Id: Ib8e1259fa0b91d29127c1675c91b4e74e46b9bfa