Azure: don't require full subnet id
The current config file for azure requires the user to supply a full
object id for the subnet.  That looks something like this:

    subnet-id: /subscriptions/........-....-....-....-............/resourceGroups/nodepoolRG/providers/Microsoft.Network/virtualNetworks/nodepool/subnets/default

It's not trivial to find this in the web portal either.  To make it
easier for users, let's deprecate this and specify it via any of
the following:

    network:
      resource-group: nodepoolRG
      network: nodepool
      subnet: default

or:

    network:
      network: nodepool

or (the most likely case):

    network: nodepool

Where resource-group will default to the resource group of the provider
and subnet will default to "default" which is what Azure initially creates.

Change-Id: I423fb6739089a44116ec00ba9b9ba219b5563fc2