Azure: add quota support
This adds support for querying the provider quota as well as
getting resource usage information for hardware profiles.

This also includes a change to the general quota utils for
all drivers:

If a driver raises an exception when querying available quota
and we have a cached value, use the cached value.  This avoids
errors from transient provider API/network issues.

Change-Id: I7199ced512226aaf8bdb6106d41796a5cbca97c0