Azure: replace driver with state machine driver
Now that the state-machine-based driver for Azure is complete,
replace the current driver with it.  This changis is generally
simply file renames, with some minor bugfixes to satisfy
edge-cases that appear in the tests.

This also updates the fake azure to accomodate the additional
methods used by the new driver, and adds a test for image uploads
and diskimage building.

Change-Id: I6b5cf72501ea83a8a7a2f753ee6ed8d2e484a5d2