Azure: fix race in leaked resources
The leaked resource detection can race with resource creation and
erroneously delete a newly created resource.  To fix this,
verify that a leaked resource shows up in two subsequent runs.

We could do this in the adapter, but if we generalize resource
representation, then we could do it once in the statemachine driver
and make adapter implementation simpler, so that's what this change
does.

Change-Id: I725bbf6901839b91781d738a7a9f07c1ebfa3369