Account for resource usage of leaked nodes
I noticed this from a callback emitted in the logs

 Traceback (most recent call last):
  File ".../nodepool/driver/utils.py", line 280, in estimatedNodepoolQuotaUsed
    if node.type[0] not in provider_pool.labels:
IndexError: list index out of range

This is caused by driver/utils.py:estimatedNodepoolQuotaUsed where it
walks the active ZK nodes and looks for those owned by this nodepool
instance and then does a calculation via the "type" to lookup the
flavor of the label, and thus the resource usage.  Clearly, we had
some nodes that did not have a "type".

It turns out the OpenStack driver puts leaked nodes into ZK with a
state of "zk.DELETING" in
driver/openstack/provider.py:cleanupLeakedInstances() with no "type"
value.  As described inline, we may not be able to lookup the
configuration of a leaked node because the config for this might have
gone away.  Thus these Zookeeper nodes don't have a "type" set, and
raise the error mentioned above.

These nodes are not accounted for in
driver/openstack/provider.py:unmanagedQuotaUsed because they are
considered managed nodes; we know they are owned by this nodepool
hence why we are deleting them.  Ergo this throws the quota
calculations out.

I think we have seen the results of this when starting in the Linaro
cloud after a period of brokenness, where we created a thundering
heard of deleting and building nodes.

The solution proposed here is to store the actual cpu/ram usage of the
running server, as returned by openstack API query of the flavor it is
running, in the zk.DELETING node.  Then on the quota side, if we do
not see a "type" in the node (giving us a key to find the flavor and
hence recourse usage in the config), we also check for the raw cpu/ram
fields and add them into the quota calculations.  This stops the
original exception and also accounts for the deleting nodes correctly.

To enable testing, we need to add the "flavor" field to the Dummy
server value that is leaked in testing.  This is part of the openstack
API for a server so should be there.  The existing "test_leaked_node"
test should then exercise this path.

Change-Id: I4a6e5af83fe4e0db2b17b0ec9b5a47f1aa10bf32