Remove statsd args to OpenStack API client call
Change If21a10c56f43a121d30aa802f2c89d31df97f121 modified nodepool to
not use the inbuilt TaskManager but use openstackapi's task handling
instead.

The statsd arguments added here don't actually do anything and are
ignored; an openstack.Connection() object doesn't setup the stats
configuration.  Things are somewhat working because of the
STATSD_<HOST|PORT> environment variables -- openstacksdk notices these
and turns on stats reporting.  However, it uses the default prefix
('openstack.api') which is a regression over the previous behaviour of
logging operations on a per-cloud basis.

I have proposed [1] that will allow setting the prefix for stats in
the "metric" section of each cloud in the openstacksdk config file.
This will allow users to return to the previous behaviour by setting
each cloud with an individual prefix in the cloud configuration (or,
indeed keep the current behaviour by not setting that).  So along with
removing the ineffective arguments, I've updated the relevant
documentation and added a release note detailing this.

https://review.opendev.org/c/openstack/openstacksdk/+/786814

Change-Id: I30e57084489d822dd6152d3e5712e3cd201372ae