nodepool-functional : ignore errors copying logs
If the job stops before the required directories here are created we
get a failure copying, which stops us collecting things that may be
useful to diagnosing the error.  Put the copying in a block that
ignores errors.

Also just copy the /var/log/syslog -- I think this came in originally
from something running on CentOS which doesn't have this, but now we
run on Ubuntu hosts and syslog has interesting things for diagnosing
errors.

Change-Id: Iaca4801a652ef4a67772c804271ea5c1db377051