QuotaInformation : abstract resource recording
Change I670d4bfd7d35ce1109b3ee9b3342fb45ee283a79 added the
quota['compute'] property to ZK nodes for Zuul to keep track of.

Instead of accessing the property directly, this adds a
get_resources() function to QuotaInformation which returns it, and
makes a point to note that the property is used so that we don't
modify it in incompatible ways in future.

This is a no-op refactor, but helpful in a follow-on change that also
adds this field to leaked nodes.

Change-Id: Id78b059cf2121e01e4cd444f6ad3834373cf7fb6