Stop installing docker via bindep
The latest docker-ce release does not start properly if it is installed
after installing and removing docker.io on Ubuntu. It is likely that the
docker.io package is leaking something across to the docker-ce install
that causes this problem. We simplify things by relying on ensure-docker
to install docker for us instead. This avoids the install, uninstall,
and broken install process.

Change-Id: I9e08dbf1ee3e6e146fb9ee6ba3435b3048096f5b