Avoid race deleting images
This came from an empty build node, for reference it was

 /nodepool/images/fedora-32/builds/0000057968

This had an empty node

 /nodepool/images/fedora-32/builds/0000057968/providers/ovh-bhs1/images

Upon investigation of the logs, I found that two builders had tried to
clean this up.

 2021-04-21 06:35:52,698 DEBUG nodepool.builder.CleanupWorker.0: Deleting image upload: <ImageUpload {'state': 'deleting', 'state_time': 1618986946.3660314, 'external_id': '517d4974-b220-47b2-8cbf-551e7d28bb88', 'external_name': 'fedora-32-1618803263', 'format': 'qcow2', 'username': 'zuul', 'python_path': '/usr/bin/python3', 'shell_type': None, 'id': '0000000001', 'build_id': '0000057968', 'provider_name': 'ovh-bhs1', 'image_name': 'fedora-32'}>

 2021-04-21 06:35:48,993 DEBUG nodepool.builder.CleanupWorker.0: Deleting image upload: <ImageUpload {'state': 'deleting', 'state_time': 1618986946.3660314, 'external_id': '517d4974-b220-47b2-8cbf-551e7d28bb88', 'external_name': 'fedora-32-1618803263', 'format': 'qcow2', 'username': 'zuul', 'python_path': '/usr/bin/python3', 'shell_type': None, 'id': '0000000001', 'build_id': '0000057968', 'provider_name': 'ovh-bhs1', 'image_name': 'fedora-32'}>

This means that both builders fell into _deleteUpload() at

   self.log.debug("Deleting image upload: %s", upload)
   try:
       with self._zk.imageUploadNumberLock(upload,
                                           blocking=False):
           self._zk.deleteUpload(upload.image_name,
                                 upload.build_id,
                                 upload.provider_name, upload.id)
   except exceptions.ZKLockException:
       # If we can't get a lock, we'll try again later.
       self.log.debug("Unable to get lock on image upload: %s",
                      upload)

We did not see an exception, so both builders must have thought they go
imageUploadNumberLock()

I think the problem here is the way that the lock node is placed under
the image node and the subsequent recursive delete.

When we go to delete "images" with deleteUpload(recursive=True) we
first delete "lock", which then indicates another process can come in
and recreate the lock, try and delete as well and we end up with nodes
left behind (I admit this last bit is a bit hand-wavy...)

The solution proposed here is to put the upload lock as a *sibling* of
the upload id.  So we have

  ...images/lock-000000001
  ...images/000000001

The trick here is that we now need to explicitly remove
"lock-000000001" when we release the upload lock.  To do this, we wrap
the existing Kazoo Lock object with a function that deletes its path.
So in the context manager, when we drop the lock, we also remove the
lock node.  By this time, the image-id is gone, so no other builder
will be trying to remove it.

Thus we do not need a recursive delete any more, this is removed.

The orphan test is modified.  It was assuming that
'imageUploadNumberLock' would create "images/<upload.id" for it, which
is now wrong.  We manually construct the path (and fix the id passing
too).

Change-Id: I7a4d7792aa4f5722504c55bfc92f7b0c19d7be42