Remove nodepool-functional-opensatck
This test installs devstack and then nodepool on a bionic host (in
contrast to the -containers variant that builds a container from the
Dockerfile and installs/runs that).

We don't really need to test if we run on a Bionic host.  We have tox
jobs testing various Python versions for compatability, so running on
Bionic isn't providing any extra coverage there.  DIB can't build many
things on plain Bionic due to updates required in RPM, Zypper,
debootstrap, etc.  The container incorporates fixes as required and is
where anyone is going to put attention if there are build issues;
hence we're not testing anything useful for image building paths.
Finally we also have nodepool-zuul-functional, which brings up Zuul
and nodepool on a plain Bionic host anyway.  Per the prior reasons,
that covers basically the same thing this is providing anyway.

The only inactive user of this is DIB in the dependent change.  It was
trying to use this for release testing -- but that's not really valid.
This test installs nodepool into a venv from the Zuul checkout; ergo
this test was not providing DIB with a release version of
nodepool-builder/launcher to test against anyway.

Depends-On: https://review.opendev.org/c/openstack/openstacksdk/+/788414
Depends-On: https://review.opendev.org/c/openstack/openstacksdk/+/788416
Depends-On: https://review.opendev.org/c/openstack/diskimage-builder/+/788404

Change-Id: I87318e9101b982f3cafcf82439fdcb68767b602b